third_party/blink/web_tests/wpt_internal/app-history/app-history-entry/key-navigate-back-cross-document.html

<!doctype html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<iframe id="i" src="resources/key-navigate-back-cross-document-helper.html"></iframe>
<script>
async_test(t => {
  window.finish = t.step_func_done((end_key, end_id) => {
    assert_equals(window.start_key, end_key);
    assert_equals(window.start_id, end_id);
  });
}, "AppHistoryEntry.key behavior on cross-document back navigation");
</script>
